Prevent FunctionalRecipeGenerator from fetching old reflections on 1.20+
